% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raxml}
\alias{raxml}
\title{raxml}
\usage{
raxml(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments to raxml provided as a character vector}
}
\description{
Run raxml
}
\examples{
library(outsider)
raxml <- module_import('raxml', repo = 'dombennett/om..raxml')
raxml('-h')
}
