% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raxml}
\alias{raxml}
\title{raxml}
\usage{
raxml(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run raxml
}
\examples{
library(outsider)
module_install(repo = 'dombennett/om..raxml', force = TRUE)
raxml <- module_import('raxml', repo = 'dombennett/om..raxml')
raxml('-h')
module_uninstall(repo = 'dombennett/om..raxml')
}
