% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raxml}
\alias{raxml}
\title{raxml}
\usage{
raxml(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run raxml
}
\examples{
library(outsider)
raxml <- module_import('raxml', repo = 'dombennett/om..raxml..8.2.12')
raxml('-h')
}
